rename dsp to promote-post for clarity
