--- client/lib/dsp/README.md
@@ -1,3 +1,0 @@
-# DSP Widget
-
-
