+++ client/lib/promote-post/README.md
@@ -1,0 +1,8 @@
+# Promote Post (DSP Widget)
+
+DSP: Demand Side Platform
+Project P2: https://wp.me/pbpazC-2Pd
+
+Allow users to promote their posts via WordPress.com and/or Tumblr.
+This is implemented by invoking our DSP Widget which handles the promote process.
+
